+++ apiphany-integration-tests/src/test/java/org/apiphany/tests/ApiClientWithJavaNetHttpIT.java
@@ -1,4 +1,4 @@
-package org.apiphany;
+package org.apiphany.tests;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.containing;
@@ -34,6 +34,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Supplier;
 import java.util.stream.IntStream;
 
+import org.apiphany.ApiClient;
+import org.apiphany.ApiRequest;
+import org.apiphany.ApiResponse;
 import org.apiphany.client.ClientProperties;
 import org.apiphany.client.DecoratingExchangeClient;
 import org.apiphany.client.ExchangeClient;